Correct rubocop issues

+++ lib/autoproj/package_managers/python.rb
@@ -6,27 +6,27 @@ module Autoproj
         # Get the python version for a given python executable
         # @return [String] The python version as <major>.<minor>
         def self.get_python_version(python_bin)
-            if !File.exist?(python_bin)
+            unless File.exist?(python_bin)
                 raise ArgumentError, "Autoproj::Python.get_python_version executable "\
                             "'#{python_bin}' does not exist"
             end
 
             cmd = "#{python_bin} -c \"import sys;"\
                 "version=sys.version_info[:3]; "\
-                "print('{0}.{1}'.format(*version))\"".strip()
+                "print('{0}.{1}'.format(*version))\"".strip
 
             msg, status = Open3.capture2e(cmd)
             if status.success?
-                python_version = msg.strip()
-                return python_version
+                msg.strip
+
             else
-                raise RuntimeError, "Autoproj::Python.get_python_version identification"\
+                raise "Autoproj::Python.get_python_version identification"\
                     " of python version for '#{python_bin}' failed: #{msg}"
             end
         end
 
         def self.get_pip_version(pip_bin)
-            if !File.exist?(pip_bin)
+            unless File.exist?(pip_bin)
                 raise ArgumentError, "Autoproj::Python.get_pip_version executable "\
                             "'#{pip_bin}' does not exist"
             end
@@ -35,20 +35,20 @@ module Autoproj
 
             msg, status = Open3.capture2e(cmd)
             if status.success?
-                pip_version = msg.split(" ")[1]
-                return pip_version
+                msg.split(" ")[1]
+
             else
-                raise RuntimeError, "Autoproj::Python.get_pip_version identification"\
+                raise "Autoproj::Python.get_pip_version identification"\
                     " of pip version for '#{pip_bin}' failed: #{msg}"
             end
         end
 
         def self.validate_version(version, version_constraint)
             if !version_constraint
-                return true
+                true
             else
                 dependency = Gem::Dependency.new("python", version_constraint)
-                return dependency.match?("python", version)
+                dependency.match?("python", version)
             end
         end
 
@@ -56,33 +56,31 @@ module Autoproj
         # a given version constraint
         # @param [String] python_bin the python executable
         # @param [String] version_constraint version constraint, e.g., <3.8, >= 3.7, 3.6
-        # @return [String,Bool] Version and validation result, i.e., True if binary fulfills the version constraint, false
-        #   otherwise
+        # @return [String,Bool] Version and validation result, i.e.,
+        #         True if binary fulfills the version constraint, false otherwise
         def self.validate_python_version(python_bin, version_constraint)
             version = get_python_version(python_bin)
-            return [version, validate_version(version, version_constraint)]
+            [version, validate_version(version, version_constraint)]
         end
 
         # Find python given a version constraint
         # @return [String,String] path to python executable and python version
         def self.find_python(ws: Autoproj.workspace,
-                             version: ws.config.get('python_version',nil))
+                             version: ws.config.get('python_version', nil))
             finders = [
-                lambda { Autobuild.programs['python'] },
-                lambda { `which python3`.strip() },
-                lambda { `which python`.strip() }
+                -> { Autobuild.programs['python'] },
+                -> { `which python3`.strip },
+                -> { `which python`.strip }
             ]
 
             finders.each do |finder|
                 python_bin = finder.call
                 if python_bin && !python_bin.empty?
                     python_version, valid = validate_python_version(python_bin, version)
-                    if valid
-                        return python_bin, python_version
-                    end
+                    return python_bin, python_version if valid
                 end
             end
-            raise RuntimeError, "Autoproj::Python.find_python_bin: failed to find python" \
+            raise "Autoproj::Python.find_python_bin: failed to find python" \
                 " for version '#{version}'"
         end
 
@@ -101,10 +99,11 @@ module Autoproj
 
             # If a version constraint is given, ensure fulfillment
             if validate_version(config_version, version)
-                return config_bin, config_version
+                [config_bin, config_version]
             else
-                raise RuntimeError, "python_executable in autoproj config with version '#{config_version}'"\
-                    " does not match version constraints '#{version}'"
+                raise "python_executable in autoproj config with " \
+                  "version '#{config_version}' does not match "\
+                  "version constraints '#{version}'"
             end
         end
 
@@ -112,9 +111,9 @@ module Autoproj
                                        version: nil)
             version, valid = validate_python_version(bin, version)
             if valid
-                return [bin, version]
+                [bin, version]
             else
-                raise RuntimeError, "Autoproj::Python.resolve_python: requested python"\
+                raise "Autoproj::Python.resolve_python: requested python"\
                     "executable '#{bin}' does not satisfy version"\
                     "constraints '#{version}'"
             end
@@ -124,30 +123,30 @@ module Autoproj
                                      version: nil)
             version_constraint = version
             resolvers = [
-                lambda { get_python_from_config(ws: ws, version: version_constraint) },
-                lambda { find_python(ws: ws, version: version_constraint) }
+                -> { get_python_from_config(ws: ws, version: version_constraint) },
+                -> { find_python(ws: ws, version: version_constraint) }
             ]
 
             bin = nil
             resolvers.each do |resolver|
-                begin
                     bin, version = resolver.call
                     if bin && File.exist?(bin) && version
-                        Autoproj.debug "Autoproj::Python.resolve_python: found python '#{bin}'"\
-                            " version '#{version}'"
+                        Autoproj.debug "Autoproj::Python.resolve_python: " \
+                          "found python '#{bin}' version '#{version}'"
                         break
                     end
-                rescue RuntimeError => e
-                    Autoproj.debug "Autoproj::Python.resolve_python: resolver failed: #{e}"
-                end
+            rescue RuntimeError => e
+                    Autoproj.debug "Autoproj::Python.resolve_python: " \
+                      "resolver failed: #{e}"
             end
 
-            if !bin
-                msg = "Autoproj::Python.resolve_python: failed to find a python executable"
+            unless bin
+                msg = "Autoproj::Python.resolve_python: " \
+                      "failed to find a python executable"
                 if version_constraint
                     msg += " satisfying version constraint '#{version_constraint}'"
                 end
-                raise RuntimeError, msg
+                raise msg
             end
             [bin, version]
         end
@@ -163,29 +162,25 @@ module Autoproj
                                 bin: nil,
                                 version: nil)
             if bin
-                return custom_resolve_python(bin: bin, version: version)
+                custom_resolve_python(bin: bin, version: version)
             else
-                return auto_resolve_python(ws: ws, version: version)
+                auto_resolve_python(ws: ws, version: version)
             end
         end
 
         def self.remove_python_shims(root_dir)
-            shim_path = File.join(root_dir, "install","bin","python")
-            if File.exist?(shim_path)
-                FileUtils.rm shim_path
-            end
+            shim_path = File.join(root_dir, "install", "bin", "python")
+            FileUtils.rm shim_path if File.exist?(shim_path)
         end
 
         def self.remove_pip_shims(root_dir)
-            shim_path = File.join(root_dir, "install","bin","pip")
-            if File.exist?(shim_path)
-                FileUtils.rm shim_path
-            end
+            shim_path = File.join(root_dir, "install", "bin", "pip")
+            FileUtils.rm shim_path if File.exist?(shim_path)
         end
 
         def self.rewrite_python_shims(python_executable, root_dir)
-            shim_path = File.join(root_dir, "install","bin")
-            if !File.exist?(shim_path)
+            shim_path = File.join(root_dir, "install", "bin")
+            unless File.exist?(shim_path)
                 FileUtils.mkdir_p shim_path
                 Autoproj.warn "Autoproj::Python.rewrite_python_shims: creating "\
                     "#{shim_path} - "\
@@ -197,13 +192,13 @@ module Autoproj
                 io.puts "#! /bin/sh"
                 io.puts "exec #{python_executable} \"$@\""
             end
-            FileUtils.chmod 0755, python_path
+            FileUtils.chmod 0o755, python_path
             python_path
         end
 
         def self.rewrite_pip_shims(python_executable, root_dir)
-            shim_path = File.join(root_dir, "install","bin")
-            if !File.exist?(shim_path)
+            shim_path = File.join(root_dir, "install", "bin")
+            unless File.exist?(shim_path)
                 FileUtils.mkdir_p shim_path
                 Autoproj.warn "Autoproj::Python.rewrite_pip_shims: creating "\
                     "#{shim_path} - "\
@@ -214,7 +209,7 @@ module Autoproj
                 io.puts "#! /bin/sh"
                 io.puts "exec #{python_executable} -m pip \"$@\""
             end
-            FileUtils.chmod 0755, pip_path
+            FileUtils.chmod 0o755, pip_path
             pip_path
         end
 
@@ -255,11 +250,11 @@ module Autoproj
                                  ws: Autoproj.workspace,
                                  bin: nil,
                                  version: nil)
-            return unless ws.config.get('USE_PYTHON',nil)
+            return unless ws.config.get('USE_PYTHON', nil)
 
             bin, version = resolve_python(ws: ws, bin: bin, version: version)
             path = File.join(pkg.prefix, "lib",
-                                 "python#{version}","site-packages")
+                             "python#{version}", "site-packages")
             pkg.env_add_path 'PYTHONPATH', path
 
             [bin, version, path]
@@ -268,51 +263,50 @@ module Autoproj
         def self.prepare_venv(ws: Autoproj.workspace, bin: nil, version: nil,
                               venv_name: ".autoproj_python_venv")
 
-          python_bin,_ = resolve_python(ws: ws, bin: bin, version: version)
+          python_bin, = resolve_python(ws: ws, bin: bin, version: version)
           cmd = "#{python_bin} -m pip install --user -U virtualenv"
           msg, status = Open3.capture2e(cmd)
-          if not status.success?
-              raise RuntimeError, "Autoproj::Python.prepare_venv installation of "\
+          unless status.success?
+              raise "Autoproj::Python.prepare_venv installation of "\
                   " virtualenv failed: '#{msg}'"
           end
 
           cmd = "#{python_bin} -m virtualenv #{venv_name}"
           msg, status = Open3.capture2e(cmd)
-          if not status.success?
-              raise RuntimeError, "Autoproj::Python.prepare_venv preparation of"\
+          unless status.success?
+              raise "Autoproj::Python.prepare_venv preparation of"\
                   " virtual env '#{venv_name}' failed: '#{msg}'"
           end
 
-          ws.env.set('VIRTUAL_ENV_DISABLE_PROMPT',1)
-          ws.env.source_after File.join(Autoproj.root_dir, venv_name, "bin","activate")
-          return File.join(Autoproj.root_dir, venv_name)
+          ws.env.set('VIRTUAL_ENV_DISABLE_PROMPT', 1)
+          ws.env.source_after File.join(Autoproj.root_dir, venv_name, "bin", "activate")
+          File.join(Autoproj.root_dir, venv_name)
         end
 
         def self.setup_python_configuration_options(ws: Autoproj.workspace)
             ws.config.declare 'USE_PYTHON', 'boolean',
-                default: 'no',
-                doc: [ "Do you want to activate python?" ]
+                              default: 'no',
+                              doc: ["Do you want to activate python?"]
 
             if ws.config.get("USE_PYTHON")
-                if !ws.config.has_value_for?('python_executable')
+                unless ws.config.has_value_for?('python_executable')
                     remove_python_shims(ws.root_dir)
                     remove_pip_shims(ws.root_dir)
-                    python_bin,_ = auto_resolve_python(ws: ws)
+                    python_bin, = auto_resolve_python(ws: ws)
                 end
 
                 ws.config.declare 'python_executable', 'string',
-                    default: "#{python_bin}",
-                    doc: [ "Select the path to the python executable" ]
+                                  default: python_bin.to_s,
+                                  doc: ["Select the path to the python executable"]
 
                 activate_python(ws: ws)
 
                 ws.config.declare 'USE_PYTHON_VENV', 'boolean',
-                    default: 'no',
-                    doc: [ "Do you want to use a virtual environment for python?" ]
+                                  default: 'no',
+                                  doc: ["Do you want to use a virtual" \
+                                        "environment for python?"]
 
-                if ws.config.get("USE_PYTHON_VENV")
-                    prepare_venv(ws: ws)
-                end
+                prepare_venv(ws: ws) if ws.config.get("USE_PYTHON_VENV")
             else
                 deactivate_python(ws: ws)
             end
