+++ test/package_managers/test_python.rb
@@ -1,6 +1,5 @@
 require 'autoproj/package_managers/python'
 require 'autoproj/test'
-#require 'flexmock'
 
 module Autoproj
     module Python
@@ -12,16 +11,14 @@ module Autoproj
                 @pkg.prefix = "/tmp/install/foo/"
                 @env = flexmock(base: Autobuild::Environment)
 
-                @test_python = File.join(Dir.tmpdir(), "test-python")
-                if !File.exist?(@test_python)
+                @test_python = File.join(Dir.tmpdir, "test-python")
+                unless File.exist?(@test_python)
                     File.symlink("/usr/bin/python",
                                  @test_python)
                 end
             end
             after do
-                if !File.exist?(@test_python)
-                    File.rm(@test_python)
-                end
+                File.rm(@test_python) unless File.exist?(@test_python)
             end
 
             it "does get the python version" do
@@ -32,7 +29,7 @@ module Autoproj
             end
 
             it "does validate the python version" do
-                version, valid = Autoproj::Python.validate_python_version("/usr/bin/python",nil)
+                version, valid = Autoproj::Python.validate_python_version("/usr/bin/python", nil)
                 assert(version =~ /[0-9]+\.[0-9]+/)
                 assert(valid)
 
@@ -50,26 +47,25 @@ module Autoproj
             end
 
             it "does find python" do
-                assert_raises { Autoproj::Python.find_python(ws: @ws, version: ">100.0" ) }
-                python_bin, version = Autoproj::Python.find_python(ws: @ws, version: "<100.0" )
+                assert_raises { Autoproj::Python.find_python(ws: @ws, version: ">100.0") }
+                python_bin, version = Autoproj::Python.find_python(ws: @ws, version: "<100.0")
                 assert(File.exist?(python_bin))
                 assert(version =~ /[0-9]+\.[0-9]+/)
 
                 # Python3 has higher priority, so should be picked
-                python_path = `which python3`.strip()
-                if File.exist?(python_path)
-                    assert(version =~ /3.[0-9]+/)
-                end
+                python_path = `which python3`.strip
+                assert(version =~ /3.[0-9]+/) if File.exist?(python_path)
             end
 
             it "custom resolve python" do
                 python_bin_resolved, _version_resolved = Autoproj::Python.custom_resolve_python(bin: @test_python)
                 assert(python_bin_resolved)
-                assert_raises { Autoproj::Python.custom_resolve_python(bin: @test_python,
-                                                           version: ">100.0") }
+                assert_raises do
+                  Autoproj::Python.custom_resolve_python(bin: @test_python,
+                                                         version: ">100.0")
+                end
 
                 assert_raises { Autoproj::Python.custom_resolve_python(bin: "no-existing-python") }
-
 
                 _python_bin, version = Autoproj::Python.find_python(ws: @ws)
                 @ws.config.set("python_executable", @test_python)
@@ -86,21 +82,20 @@ module Autoproj
             end
 
             it "does update the python path" do
-                @ws.config.set("USE_PYTHON",true)
+                @ws.config.set("USE_PYTHON", true)
                 bin, version, sitelib_path = Autoproj::Python.activate_python_path(@pkg, ws: @ws)
 
                 python_bin_name = File.basename(bin)
-                python_bin = `which #{python_bin_name}`.strip()
-                assert($? == 0, "This test requires python to be available on your"\
+                python_bin = `which #{python_bin_name}`.strip
+                assert($CHILD_STATUS == 0, "This test requires python to be available on your"\
                        " system, so please install before running this test")
-
 
                 assert(python_bin == bin, "Python bin #{python_bin} not equal to #{bin}")
                 assert(version == Autoproj::Python.get_python_version(python_bin))
-                assert(sitelib_path == File.join(@pkg.prefix, "lib","python#{version}","site-packages"))
+                assert(sitelib_path == File.join(@pkg.prefix, "lib", "python#{version}", "site-packages"))
 
                 found_path = false
-                path_pattern = File.join(@pkg.prefix,"lib","python.*","site-packages")
+                path_pattern = File.join(@pkg.prefix, "lib", "python.*", "site-packages")
 
                 @env.should_receive(:add_path)
                 op = @pkg.apply_env(@env).first
@@ -108,9 +103,7 @@ module Autoproj
                 assert(op.type == :add_path)
                 assert(op.name == "PYTHONPATH")
                 op.values.each do |p|
-                    if p =~ /#{path_pattern}/
-                        found_path = true
-                    end
+                    found_path = true if p =~ /#{path_pattern}/
                 end
                 assert(found_path)
                 assert(!@ws.config.has_value_for?('python_executable'))
@@ -125,16 +118,16 @@ module Autoproj
 
             it "does not update python path" do
                 @ws.config.reset
-                @ws.config.set('interactive',false)
-                @ws.config.set('USE_PYTHON',false)
+                @ws.config.set('interactive', false)
+                @ws.config.set('USE_PYTHON', false)
 
                 pkg = flexmock('testpkg')
-                prefix = File.join(@ws.root_dir,"install","testpkg")
+                prefix = File.join(@ws.root_dir, "install", "testpkg")
                 pkg.should_receive(:prefix).and_return(prefix)
                 assert(!@ws.config.has_value_for?('python_executable'))
                 assert(!@ws.config.has_value_for?('python_version'))
 
-                bin,version,path = Autoproj::Python.activate_python_path(pkg)
+                bin, version, path = Autoproj::Python.activate_python_path(pkg)
                 assert(!(bin || version || path))
             end
 
@@ -143,12 +136,12 @@ module Autoproj
                 assert(@ws.config.has_value_for?('python_executable'))
                 assert(@ws.config.has_value_for?('python_version'))
 
-                python_bin = File.join(@ws.root_dir, "install","bin","python")
+                python_bin = File.join(@ws.root_dir, "install", "bin", "python")
                 assert(File.exist?(python_bin))
                 python_version = Autoproj::Python.get_python_version(python_bin)
                 assert(python_version == @ws.config.get('python_version'))
 
-                pip_bin = File.join(@ws.root_dir, "install","bin","pip")
+                pip_bin = File.join(@ws.root_dir, "install", "bin", "pip")
                 assert(File.exist?(pip_bin))
                 pip_version = Autoproj::Python.get_pip_version(pip_bin)
                 expected_pip_version = `#{python_bin} -c "import pip; print(pip.__version__)"`.strip
@@ -157,15 +150,15 @@ module Autoproj
 
             it "does setup python" do
                 @ws.config.reset
-                @ws.config.set('interactive',false)
-                @ws.config.set('USE_PYTHON',true)
+                @ws.config.set('interactive', false)
+                @ws.config.set('USE_PYTHON', true)
                 Autoproj::Python.setup_python_configuration_options(ws: @ws)
                 assert(@ws.config.get('USE_PYTHON'))
                 assert(@ws.config.get('python_executable'))
                 assert(@ws.config.get('python_version'))
 
                 @ws.config.reset
-                @ws.config.set('interactive',false)
+                @ws.config.set('interactive', false)
                 Autoproj::Python.setup_python_configuration_options(ws: @ws)
                 if Autoproj::VERSION > '2.11.0'
                     assert(!@ws.config.get('USE_PYTHON'))
